web_bb_project/web_bb/requirements.txt
Flask==3.1.0
flask_sqlalchemy==3.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.33
pymysql
cryptography
gunicorn
uwsgi